TiliquaSoc generic, selftest only
